Keep generation provider readiness honest
Anthropic generation support is available in code, but the public status should not imply a configured deployment until a runtime has a real API key and end-to-end validation. The provider-specific implementation stays outside the model-agnostic generator boundary, and local Spec Kitty dashboard state is ignored rather than committed.

Constraint: Feature 006 remains placeholder-readiness until deployment configuration exists
Rejected: Mark generation_provider as configured | no deployment secret wiring or live validation is visible in this repo
Confidence: high
Scope-risk: narrow
Directive: Do not mark the generation provider configured/validated without deployment evidence and a real provider validation path
Tested: npm run typecheck; npm test; node scripts/verify-status-consistency.mjs; targeted Anthropic provider tests
Not-tested: Live Anthropic API call with a deployed ANTHROPIC_API_KEY
